+++ .gitignore
@@ -1,5 +1,5 @@
 venv/
-tests/
+./src/tests/
 __pycache__/
 .vscode/
 build/
